Rename a few things, last commit before I revert all the things to function components again because *apparently* somebody at React decided those were better than class components and deprecated the latter.
