+++ src/components/DegreeQualifications/DegreeQualifications.css
